Extract Python-specific SugaredValues to a separate file from init.cpp.
